fixed type of strlen() return in fprintf usage
git-svn-id: https://svn.code.sf.net/p/teem/code/teem/trunk@1931 3d70eeeb-363e-0410-a505-8a46323a89f2

+++ src/nrrd/test/kv.c
@@ -48,7 +48,8 @@ main(int argc, char **argv) {
   }
 
   if ((value = nrrdKeyValueGet(nrrd, key))) {
-    fprintf(stderr, "%s: '%s':='%s' (%d)\n", me, key, value, strlen(value));
+    fprintf(stderr, "%s: '%s':='%s' (%d)\n", me, key, value, 
+	    (int)strlen(value));
   } else {
     fprintf(stderr, "%s: value not found for key: %s\n", me, key);
   }
